add generic type and move code to new file
