+++ wandb/util.py
@@ -1,4 +1,3 @@
-import asyncio
 import colorsys
 import contextlib
 import dataclasses
@@ -38,7 +37,6 @@ from typing import (
     IO,
     TYPE_CHECKING,
     Any,
-    Awaitable,
     Callable,
     Dict,
     Generator,
@@ -66,7 +64,7 @@ from wandb.errors import (
     term,
 )
 from wandb.sdk.internal.thread_local_settings import _thread_local_api_settings
-from wandb.sdk.lib import asyncio_compat, filesystem, printer, runid
+from wandb.sdk.lib import filesystem, runid
 from wandb.sdk.lib.json_util import dump, dumps
 from wandb.sdk.lib.paths import FilePathStr, StrPath
 
@@ -1997,45 +1995,3 @@ class NonOctalStringDumper(yaml.Dumper):
         if tag == "tag:yaml.org,2002:str" and value.startswith("0") and len(value) > 1:
             return super().represent_scalar(tag, value, style="'")
         return super().represent_scalar(tag, value, style)
-
-
-def run_async_with_spinner(
-    text: str,
-    func: Callable[[], Awaitable[Any]],
-) -> None:
-    """Run an async function and display a loading icon while it runs.
-
-    Args:
-        text: The text to display next to the spinner, while the function runs.
-        func: The function to run.
-
-    Returns:
-        The result of func.
-    """
-    spinner_printer = printer.new_printer()
-
-    async def _loop_run_with_spinner(
-        text: str,
-        func: Callable[[], Awaitable[any]],
-    ) -> None:
-        func_running = asyncio.Event()
-
-        async def update_spinner() -> None:
-            tick = 0
-            with spinner_printer.dynamic_text() as text_area:
-                if text_area:
-                    while not func_running.is_set():
-                        spinner = spinner_printer.loading_symbol(tick)
-                        text_area.set_text(f"{spinner} {text}...")
-                        tick += 1
-                        await asyncio.sleep(0.1)
-                else:
-                    spinner_printer.display(text)
-
-        async with asyncio_compat.open_task_group() as group:
-            group.start_soon(update_spinner())
-            res = await asyncio.get_running_loop().run_in_executor(None, func)
-            func_running.set()
-            return res
-
-    return asyncio_compat.run(functools.partial(_loop_run_with_spinner, text, func))
